% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{Categorical}
\alias{Categorical}
\title{Initialize a Categorical likelihood node}
\usage{
Categorical(K = NULL, input_dim = NULL)
}
\arguments{
\item{K}{an integer specifying the number of classes in the training data. If \code{K} is set to \code{NULL} (the default),
its value will be inferred from the training data when running \code{\link[=dgp]{dgp()}}. Defaults to \code{NULL}.}

\item{input_dim}{a vector that specifies which GP node outputs feed into this likelihood node. It should be either:
\itemize{
\item A vector of length \code{1} (for binary output), or
\item A vector of length \code{K} (for \code{K}-class output), containing the indices of one or \code{K} GP nodes in the feeding layer.
}

If \code{input_dim} is set to \code{NULL} (the default), all GP node outputs from the feeding layer feed into the likelihood node.
In this case, you must ensure that the feeding layer contains exactly one or \code{K} GP nodes. Defaults to \code{NULL}.}
}
\value{
A 'python' object to represent a categorical likelihood node.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#new}{\figure{lifecycle-new.svg}{options: alt='[New]'}}}{\strong{[New]}}

This function constructs a likelihood object to represent a categorical likelihood node.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\note{
The categorical likelihood node can only be linked to one or \code{K} feeding GP nodes.
}
\examples{
\dontrun{

# Check https://mingdeyu.github.io/dgpsi-R/ for examples
# on how to customize DGP structures using Categorical().
}
}
