% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{container}
\alias{container}
\title{Initialize the container object}
\usage{
container(structure, local_input_idx = NULL)
}
\arguments{
\item{structure}{a list that contains the trained structure of GP or DGP of a computer model. For GP,
this is the list exported from \code{\link[=export]{export()}}. For DGP, this is the list exported from \code{\link[=estimate]{estimate()}} .}

\item{local_input_idx}{a vector that specifies the indices (i.e., the columns) of outputs (a matrix)
produced by all models in the feeding layer that are input to the emulator
represented by the \code{structure} argument. The indices should be ordered in such a way that the extracted
output from the feeding layer is sorted in the same order as the training input used for the GP/DGP
emulator that the \code{structure} argument represents. When the emulator is in the first
layer, \code{local_input_idx} gives the indices of its input in the global testing input set, see \code{\link[=predict]{predict()}}
for descriptions of the global testing input set. Defaults to \code{NULL}. When the
argument is \code{NULL}, one needs to set its value by applying \code{\link[=set_local_input]{set_local_input()}} to the returned container object.}
}
\value{
A container object to be used for linked GP emulation.
}
\description{
This function constructs an object that contains the trained GP or DGP emulator of a computer model
for linked (D)GP emulation.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
