% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\alias{draw.gp}
\alias{draw.dgp}
\alias{draw.bundle}
\title{Validation plots of a sequential design}
\usage{
draw(object, type, ...)

\method{draw}{gp}(object, type = "rmse", ...)

\method{draw}{dgp}(object, type = "rmse", ...)

\method{draw}{bundle}(object, type = "rmse", ...)
}
\arguments{
\item{object}{can be one of the following emulator classes:
\itemize{
\item the S3 class \code{gp}.
\item the S3 class \code{dgp}.
\item the S3 class \code{bundle}.
}}

\item{type}{either \code{rmse}, for the trace plot of RMSEs of emulators constructed during the sequential designs,
or \code{design}, for visualizations of input designs created by the sequential design procedure. Defaults to \code{rmse}.}

\item{...}{N/A.}
}
\value{
A \code{patchwork} object.
}
\description{
This function draws validation plots of the sequential design of a (D)GP emulator or a bundle of (D)GP emulators.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\examples{
\dontrun{

# See design() for an example.
}
}
