% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{emulator}
\alias{emulator}
\title{Initialize the emulator object}
\usage{
emulator(all_layer, N = 50, nb_parallel = FALSE)
}
\arguments{
\item{all_layer}{the trained DGP model produced by \code{\link[=estimate]{estimate()}}.}

\item{N}{the number of imputation to produce the predictions. Increase the value to account for
more imputation uncertainties. Defaults to \code{50}.}

\item{nb_parallel}{whether to use the multi-threading to accelerate the predictions. Defaults to \code{FALSE}.}
}
\value{
An emulator object to be used by \code{\link[=predict]{predict()}} or \code{\link[=ppredict]{ppredict()}} for DGP predictions.
}
\description{
This function constructs an object for deep Gaussian process predictions.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R}.
}
