% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{estimate}
\alias{estimate}
\title{Estimating the final DGP model}
\usage{
estimate(obj, burnin = NULL)
}
\arguments{
\item{obj}{the DGP object produced by \code{\link[=train]{train()}} function.}

\item{burnin}{the number of SEM iterations to be discarded for
point estimate calculation. Must be smaller than the SEM iterations
implemented. If this is not specified, only the last 25\% of iterations
are used. Defaults to \code{NULL}.}
}
\value{
a DGP object to be used to construct the emulator by \code{\link[=emulator]{emulator()}} function.
}
\description{
This function estimates the final DGP model from the model trained using \code{\link[=train]{train()}} function.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
