% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{export}
\alias{export}
\title{Export the trained GP model}
\usage{
export(obj)
}
\arguments{
\item{obj}{the GP model produced by \code{\link[=train]{train()}} function.}
}
\value{
A GP object representing the trained GP.
}
\description{
This function exports the GP model trained using \code{\link[=train]{train()}} to be used for
linked Gaussian process constructions.
}
\details{
See examples in tutorials at \url{https://mingdeyu.github.io/dgpsi-R}.
}
