% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_thread_num}
\alias{get_thread_num}
\title{Get the number of threads}
\usage{
get_thread_num()
}
\value{
the number of threads.
}
\description{
This function gets the number of threads used for parallel computations involved
in the package.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
