% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Initialize the GP object}
\usage{
gp(X, Y, kernel)
}
\arguments{
\item{X}{a matrix where each row is an input data point and each column is an input dimension.}

\item{Y}{a matrix with only one column and each row being an input data point.}

\item{kernel}{a kernel object produced by the \code{\link[=kernel]{kernel()}} function.}
}
\value{
A GP object to be used by \code{\link[=train]{train()}} for GP training.
}
\description{
This function constructs an object for Gaussian process training.
}
