% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{Hetero}
\alias{Hetero}
\title{Initialize the heteroskedastic Gaussian likelihood}
\usage{
Hetero(input_dim = NULL)
}
\arguments{
\item{input_dim}{a vector of length two that contains the indices of two GPs in the feeding
layer whose outputs feed into the likelihood node. When set to \code{NULL},
all outputs from GPs of feeding layer feed into the likelihood node, and in this case
one needs to ensure there are only two GP nodes specified in the feeding layer..
Defaults to \code{NULL}.}
}
\value{
An object to represent the heteroskedastic Gaussian likelihood node.
}
\description{
This function constructs an object for heteroskedastic Gaussian likelihood.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R}.
}
\note{
The heteroskedastic Gaussian likelihood node only needs two feeding GP nodes.
}
