% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{NegBin}
\alias{NegBin}
\title{Initialize a negative Binomial likelihood node}
\usage{
NegBin(input_dim = NULL)
}
\arguments{
\item{input_dim}{a vector of length two that contains the indices of two GP nodes in the feeding
layer whose outputs feed into this likelihood node. When set to \code{NULL},
all outputs from GP nodes in the feeding layer feed into this likelihood node, and in such a case
one needs to ensure that only two GP nodes are specified in the feeding layer.
Defaults to \code{NULL}.}
}
\value{
A 'python' object to represent a negative Binomial likelihood node.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated and will be removed in the next release.
To incorporate a negative Binomial likelihood node into a DGP structure,
use the \code{likelihood} argument in the \code{dgp()} function instead.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\note{
The negative Binomial likelihood node can only be linked to two feeding GP nodes.
}
\examples{
\dontrun{

# Check https://mingdeyu.github.io/dgpsi-R/ for examples
# on how to customize DGP structures using NegBin().
}
}
\keyword{internal}
