% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{Poisson}
\alias{Poisson}
\title{Initialize the Poisson likelihood}
\usage{
Poisson(input_dim = NULL)
}
\arguments{
\item{input_dim}{a vector of length one that contains the indices of one GP in the feeding
layer whose outputs feed into the likelihood node. When set to \code{NULL},
all outputs from GPs of the feeding layer feed into the likelihood node, and in this case
one needs to ensure there is only one GP node specified in the feeding layer..
Defaults to \code{NULL}.}
}
\value{
An object to represent the Poissonlikelihood node.
}
\description{
This function constructs an object for Poisson likelihood.
}
\details{
See examples in tutorials at \url{https://mingdeyu.github.io/dgpsi-R}.
}
\note{
The Poisson likelihood node only needs one feeding GP node.
}
