% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{ppredict}
\alias{ppredict}
\title{Multi-core predictions from GP, DGP, or linked (D)GP emulators}
\usage{
ppredict(
  obj,
  x,
  method = "mean_var",
  full_layer = FALSE,
  sample_size = 50,
  chunk_num = NULL,
  core_num = NULL
)
}
\arguments{
\item{obj}{same as that in \code{\link[=predict]{predict()}}.}

\item{x}{same as that in \code{\link[=predict]{predict()}}.}

\item{method}{same as that in \code{\link[=predict]{predict()}}.}

\item{full_layer}{same as that in \code{\link[=predict]{predict()}}.}

\item{sample_size}{same as that in \code{\link[=predict]{predict()}}.}

\item{chunk_num}{the number of chunks that the testing input matrix \code{x} will be divided into.
Defaults to \code{NULL}. If not specified, the number of chunks is set to \code{core_num}.}

\item{core_num}{the number of cores/workers to be used. Defaults to \code{NULL}. If not specified,
the number of cores is set to \verb{(max physical cores available - 1)}.}
}
\value{
Same as that in \code{\link[=predict]{predict()}}.
}
\description{
This function implements multi-core prediction from GP, DGP, or linked (D)GP emulators.
}
\details{
See examples in tutorials at \url{https://mingdeyu.github.io/dgpsi-R}.
}
