% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read}
\alias{read}
\title{Load the stored emulator object}
\usage{
read(pkl_file)
}
\arguments{
\item{pkl_file}{the path to and the name of the \code{.pkl} file where the emulator is stored.}
}
\value{
An emulator object. See \code{\link[=write]{write()}}.
}
\description{
This function loads the \code{.pkl} file that stores the emulator.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R}.
}
