% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{set_local_input}
\alias{set_local_input}
\title{Set local input indices}
\usage{
set_local_input(obj, idx)
}
\arguments{
\item{obj}{the container object produced by \code{\link[=container]{container()}}.}

\item{idx}{same as \code{local_input_idx} in \code{\link[=container]{container()}}.}
}
\value{
A container object to be used for linked GP emulation.
}
\description{
This function sets the value of \code{local_input_idx} to a container object if the value is not set
when the container object is constructed by \code{\link[=container]{container()}}.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R}.
}
\note{
This function is useful when different models are emulated by different teams. Each team can create the container
of their model even without knowing how different models are connected together. When this information is available and
containers of different emulators are collected, the connections between emulators can then be set by assigning
values to \code{local_input_idx} of each container with this function.
}
