% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_nb_parallel}
\alias{set_nb_parallel}
\title{Implement multi-threading in predictions}
\usage{
set_nb_parallel(obj, nb_parallel)
}
\arguments{
\item{obj}{the object produced by \code{\link[=emulator]{emulator()}} or \code{\link[=lgp]{lgp()}}}

\item{nb_parallel}{\code{TRUE} to switch on Numba's multi-threading implementation and \code{FALSE} to switch off.}
}
\value{
An object for DGP or linked (D)GP with Numba's multi-threading implementation switched on or off.
}
\description{
This function switch on or off Numba's multi-threading implementation for DGP and linked (D)GP predictions.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R}.
}
\note{
One can switch on or off Numba's multi-threading implementation for prediction while constructing the DGP and linked (D)GP
objectives through \code{\link[=emulator]{emulator()}} and \code{\link[=lgp]{lgp()}}. This function is useful when one wants to change the implementation for prediction
after the DGP and linked (D)GP objectives have already been built.
}
