% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_seed}
\alias{set_seed}
\title{Random seed generator}
\usage{
set_seed(seed)
}
\arguments{
\item{seed}{a single integer value.}
}
\value{
No return value.
}
\description{
This function initializes a random number generator that sets the random seed in both R and Python
to ensure reproducible results from the package.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\examples{
\dontrun{

# See dgp() for an example.
}
}
