% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_thread_num}
\alias{set_thread_num}
\title{Set the number of threads}
\usage{
set_thread_num(num)
}
\arguments{
\item{num}{the number of threads. If it is greater than the maximum number of threads available, the
number of threads will be set to the maximum value.}
}
\value{
No return value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#new}{\figure{lifecycle-new.svg}{options: alt='[New]'}}}{\strong{[New]}}

This function sets the number of threads for parallel computations involved
in the package.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
