% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trace_plot}
\alias{trace_plot}
\title{Plot of DGP model parameter traces}
\usage{
trace_plot(
  obj,
  layer_no,
  ker_no,
  width = 4,
  height = 1,
  ticksize = 5,
  labelsize = 8,
  hspace = 0.1
)
}
\arguments{
\item{obj}{The DGP object produced by \code{\link[=train]{train()}} function.}

\item{layer_no}{the index of the interested layer.}

\item{ker_no}{the index of the interested GP in the layer specified by \code{layer_no}.}

\item{width}{the overall plot width. Defaults to \code{4}.}

\item{height}{the overall plot height. Defaults to \code{1}.}

\item{ticksize}{the size of sub-plot ticks. Defaults to \code{5}.}

\item{labelsize}{the font size of y labels. Defaults to \code{8}.}

\item{hspace}{the space between sub-plots. Defaults to \code{0.1}.}
}
\description{
This function plots the traces of model parameters of a particular GP node
in the trained DGP hierarchy.
}
\details{
See examples in tutorials at \url{https://mingdeyu.github.io/dgpsi-R}.
}
