% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train}
\alias{train}
\title{Training of GP and DGP}
\usage{
train(obj, N = 500, ess_burn = 10, disable = FALSE)
}
\arguments{
\item{obj}{a GP or DGP model produced by \code{\link[=gp]{gp()}} or \code{\link[=dgp]{dgp()}} function.}

\item{N}{number of iterations for stochastic EM. Defaults to \code{500}. Only used if \code{obj} is a DGP object.}

\item{ess_burn}{number of burnin steps for the ESS-within-Gibbs
at each I-step of the SEM. Defaults to \code{10}. Only used if \code{obj} is a DGP object.}

\item{disable}{(bool, optional): whether to disable the training progress bar.
Defaults to \code{FALSE}. Only used if \code{obj} is a DGP object.}
}
\value{
Updated GP or DGP object to be used for GP or DGP predictions.
}
\description{
This function trains a GP or DGP model constructed by \code{\link[=gp]{gp()}} or \code{\link[=dgp]{dgp()}} function.
}
\details{
See examples in tutorials at \url{https://mingdeyu.github.io/dgpsi-R}.
}
\note{
The function can be re-applied to the returned trained DGP object to continue
the training of the DGP model with additional number of SEM iterations.
}
