% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine layers}
\usage{
combine(...)
}
\arguments{
\item{...}{a sequence of lists, each of which contains the GP nodes (produced by \code{\link[=kernel]{kernel()}}),
likelihood nodes (e.g., produced by \code{\link[=Poisson]{Poisson()}}), or containers (produced by \code{\link[=container]{container()}})
in that layer.}
}
\value{
A list of layers defining the DGP or linked (D)GP structure.
}
\description{
This function combines layers into one list as a DGP or linked (D)GP structure.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
