% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine layers}
\usage{
combine(...)
}
\arguments{
\item{...}{a sequence of lists:
\enumerate{
\item For DGP emulations, each list represents a DGP layer and contains GP nodes (produced by \code{\link[=kernel]{kernel()}}), or
likelihood nodes (produced by \code{\link[=Poisson]{Poisson()}}, \code{\link[=Hetero]{Hetero()}}, or \code{\link[=NegBin]{NegBin()}}).
\item For linked (D)GP emulations, each list represents a system layer and contains emulators (produced by \code{\link[=gp]{gp()}} or
\code{\link[=dgp]{dgp()}}) in that layer.
}}
}
\value{
A list defining a DGP structure (for \code{struc} of \code{\link[=dgp]{dgp()}}) or a linked (D)GP structure
(for \code{struc} for \code{\link[=lgp]{lgp()}}).
}
\description{
This function combines customized layers into a DGP or linked (D)GP structure.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\examples{
\dontrun{

# See lgp() for an example.
}
}
