% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine layers}
\usage{
combine(...)
}
\arguments{
\item{...}{a sequence of lists. Each list represents a system layer and contains emulators (produced by \code{\link[=gp]{gp()}} or
\code{\link[=dgp]{dgp()}}) in that layer.}
}
\value{
A list defining a linked (D)GP structure to be passed to \code{struc} of \code{\link[=lgp]{lgp()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated and will be removed in the next release, as it is
simply a wrapper for the \code{\link[=list]{list()}} function. To construct linked (D)GP structures,
please use the updated \code{\link[=lgp]{lgp()}} function, which provides a simpler and more efficient
approach to building (D)GP emulators.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\keyword{internal}
