% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Initialize the DGP object}
\usage{
dgp(X, Y, all_layer = NULL, check_rep = TRUE, rff = FALSE, M = NULL)
}
\arguments{
\item{X}{a matrix where each row is an input training data point and each column is an input dimension.}

\item{Y}{a matrix containing observed training output data..
The matrix has it rows being output data points and columns being output dimensions
(with the number of columns equals to the number of GP nodes in the final layer).}

\item{all_layer}{a list contains \emph{L} (the number of layers) sub-lists, each of which contains
the GPs defined by the \code{\link[=kernel]{kernel()}} function in that layer. The sub-lists are placed in the list
in the same order of the specified DGP model. The final layer of DGP hierarchy can be set to a likelihood
layer by putting an object created by a likelihood function (e.g., \code{\link[=Poisson]{Poisson()}}) into the final sub-list of \code{all_layer}.
Defaults to \code{NULL}. If a DGP structure is not provided, an input-connected two-layered DGP structure
(for deterministic model emulation) with the number of GP nodes in the first layer equal to the dimension
of \code{X} is automatically constructed.}

\item{check_rep}{whether to check the repetitions in the dataset, i.e., if one input
position has multiple outputs. Defaults to \code{TRUE}.}

\item{rff}{whether to use random Fourier features to approximate the correlation matrices
during the imputation in training. Defaults to \code{FALSE}.}

\item{M}{the number of features to be used by random Fourier approximation. It is only used
when \code{rff} is set to \code{TRUE}. Defaults to \code{NULL}. If it is not specified, \code{M} is set to
\verb{max(100, ceil(sqrt(Data Size)*log(Data Size))))}.}
}
\value{
A DGP object to be used by \code{\link[=train]{train()}} for DGP training.
}
\description{
This function constructs an object for deep Gaussian process training.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R}.
}
