% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgpsi-package.R
\docType{package}
\name{dgpsi-package}
\alias{dgpsi}
\alias{dgpsi-package}
\title{dgpsi: Interface to 'dgpsi' for Deep and Linked Gaussian Process Emulations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Interface to the 'python' package 'dgpsi' for Gaussian process, deep Gaussian process, and linked deep Gaussian process emulations of computer models and networks using stochastic imputation (SI). The implementations follow Ming & Guillas (2021) \doi{10.1137/20M1323771} and Ming, Williamson, & Guillas (2023) \doi{10.1080/00401706.2022.2124311} and Ming & Williamson (2023) \doi{10.48550/arXiv.2306.01212}. To get started with the package, see \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mingdeyu/dgpsi-R}
  \item \url{https://mingdeyu.github.io/dgpsi-R/}
  \item Report bugs at \url{https://github.com/mingdeyu/dgpsi-R/issues}
}

}
\author{
\strong{Maintainer}: Deyu Ming \email{deyu.ming.16@ucl.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Daniel Williamson \email{d.williamson@exeter.ac.uk}
}

}
\keyword{internal}
