% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Gaussian process emulator construction}
\usage{
gp(
  X,
  Y,
  struc = NULL,
  name = "sexp",
  lengthscale = rep(0.2, ncol(X)),
  nugget_est = FALSE,
  nugget = 1e-06,
  training = TRUE,
  verb = TRUE,
  internal_input_idx = NULL,
  linked_idx = NULL
)
}
\arguments{
\item{X}{a matrix where each row is an input data point and each column is an input dimension.}

\item{Y}{a matrix with only one column and each row being an output data point.}

\item{struc}{an object produced by \code{\link[=kernel]{kernel()}} that gives a user-defined GP specifications. When \code{struc = NULL},
the GP specifications are automatically generated using information provided in \code{name}, \code{lengthscale},
\code{nugget_est}, \code{nugget}, and \code{internal_input_idx}. Defaults to \code{NULL}.}

\item{name}{kernel function to be used. Either \code{"sexp"} for squared exponential kernel or
\code{"matern2.5"} for Matérn-2.5 kernel. Defaults to \code{"sexp"}. This argument is only used when \code{struc = NULL}.}

\item{lengthscale}{initial values of lengthscales in the kernel function. It can be a single numeric value or a vector:
\itemize{
\item if it is a single numeric value, it is assumed that kernel functions across input dimensions share the same lengthscale;
\item if it is a vector (which must have a length of \code{ncol(X)}), it is assumed that kernel functions across input dimensions have different lengthscales.
}

Defaults to a vector of 0.2. This argument is only used when \code{struc = NULL}.}

\item{nugget_est}{a bool indicating if the nugget term is to be estimated:
\enumerate{
\item \code{FALSE}: the nugget term is fixed to \code{nugget}.
\item \code{TRUE}: the nugget term will be estimated.
}

Defaults to \code{FALSE}. This argument is only used when \code{struc = NULL}.}

\item{nugget}{the initial nugget value. If \code{nugget_est = FALSE}, the assigned value is fixed during the training.
Set \code{nugget} to a small value (e.g., \code{1e-6}) and the corresponding bool in \code{nugget_est} to \code{FASLE} for deterministic emulations where the emulator
interpolates the training data points. Set \code{nugget} to a reasonable larger value and the corresponding bool in \code{nugget_est} to \code{TRUE} for stochastic
emulations where the computer model outputs are assumed to follow a homogeneous Gaussian distribution. Defaults to \code{1e-6}. This argument is only used
when \code{struc = NULL}.}

\item{training}{a bool indicating if the initialized GP emulator will be trained.
When set to \code{FALSE}, \code{\link[=gp]{gp()}} returns an untrained GP emulator, to which one can apply \code{\link[=summary]{summary()}} to inspect its specifications
(especially when a customized \code{struc} is provided) or apply \code{\link[=predict]{predict()}} to check its emulation performance before the training. Defaults to \code{TRUE}.}

\item{verb}{a bool indicating if the trace information on GP emulator construction and training will be printed during the function execution.
Defaults to \code{TRUE}.}

\item{internal_input_idx}{the column indices of \code{X} that are generated by the linked emulators in the feeding layer.
Set \code{internal_input_idx = NULL} if the GP emulator is not linked to any emulator (e.g., the GP emulator is in the first layer of a system) or all columns in \code{X} are
generated by the linked emulators in the feeding layer. Defaults to \code{NULL}. This argument is only used when \code{struc = NULL}.}

\item{linked_idx}{the indices of columns in the pooled output matrix (formed by column-combined outputs of all emulators
in the feeding layer) that will feed into the GP emulator. The length of \code{linked_idx} shall equal to the length of \code{internal_input_idx}
when \code{internal_input_idx} is not \code{NULL}. Set \code{linked_idx = NULL} if the GP emulator is not intended
for linked emulations. If the GP emulator is in the first layer of a system, \code{linked_idx} gives the column indices of the
global input (formed by column-combining all input matrices of emulators in the first layer) that the GP emulator will use.
Defaults to \code{NULL}.}
}
\value{
An S3 class named \code{gp} that can be used by
\itemize{
\item \code{\link[=predict]{predict()}} for GP predictions.
\item \code{\link[=lgp]{lgp()}} to construct linked (D)GP emulators.
}
}
\description{
This function builds and trains a GP emulator.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\note{
Any R vector detected in \code{X} and \code{Y} will be treated as a column vector and automatically converted into a single-column
R matrix.
}
