% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp.R
\name{lgp}
\alias{lgp}
\title{Initialize the linked (D)GP object}
\usage{
lgp(all_layer, N = 50, nb_parallel = FALSE)
}
\arguments{
\item{all_layer}{a list contains \emph{L} (the number of layers of a systems of computer models) sub-lists,
each of which represents a layer and contains the GP/DGP emulators of computer models represented by
containers constructed by \code{\link[=container]{container()}}. The sub-lists are placed in the list in the same order of
the specified computer model system.}

\item{N}{the number of imputation to produce the predictions. Increase the value to account for more
imputation uncertainties. If the system consists only GP emulators, \code{N} is set to \code{1} automatically.
Defaults to \code{50}.}

\item{nb_parallel}{whether to use the multi-threading to accelerate the predictions. Defaults to \code{FALSE}.}
}
\value{
A linked (D)GP object to be used by \code{\link[=predict]{predict()}} or \code{\link[=ppredict]{ppredict()}} for linked (D)GP predictions.
}
\description{
This function constructs an object for linked (D)GP predictions.
}
