% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{nllik}
\alias{nllik}
\title{Calculate negative predicted log-likelihood}
\usage{
nllik(obj, x, y)
}
\arguments{
\item{obj}{a DGP object produced by \code{\link[=emulator]{emulator()}}}

\item{x}{a matrix where each row is an input testing data point and each column is an input dimension.}

\item{y}{a matrix with only one column where each row is a scalar-valued testing output data point.}
}
\value{
A named list with two components. The first one, named \code{meanNLL}, is a scalar that gives the average negative
predicted log-likelihood across all testing data points. The second one, named \code{NLL}, is a vector that gives
the negative predicted log-likelihood for each testing data point.
}
\description{
This function compute the negative predicted log-likelihood from a
trained DGP with likelihood layer.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R}.
}
