% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{Poisson}
\alias{Poisson}
\title{Initialize a Poisson likelihood node}
\usage{
Poisson(input_dim = NULL)
}
\arguments{
\item{input_dim}{a vector of length one that contains the indices of one GP node in the feeding
layer whose outputs feed into this likelihood node. When set to \code{NULL},
all outputs from GP nodes in the feeding layer feed into this likelihood node, and in such a case
one needs to ensure that only one GP node is specified in the feeding layer.
Defaults to \code{NULL}.}
}
\value{
A 'python' object to represent a Poisson likelihood node.
}
\description{
This function constructs a likelihood object to represent a Poisson likelihood node.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\note{
The Poisson likelihood node can only be linked to one feeding GP node.
}
\examples{
\dontrun{

# Check https://mingdeyu.github.io/dgpsi-R/ for examples
# on how to customize DGP structures using Poisson().
}
}
