% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict}
\alias{predict}
\title{Predictions from GP, DGP, or linked (D)GP emulators}
\usage{
predict(obj, x, method = "mean_var", full_layer = FALSE, sample_size = 50)
}
\arguments{
\item{obj}{a GP, DGP, or linked (D)GP object produced by \code{\link[=gp]{gp()}} (after applying \code{\link[=train]{train()}}), \code{\link[=emulator]{emulator()}}, or \code{\link[=lgp]{lgp()}}.}

\item{x}{a matrix where each row is an input testing data point and each column is an input dimension.}

\item{method}{the prediction approach: mean-variance (\code{"mean_var"}) or sampling (\code{"sampling"}) approach. Defaults to \code{"mean_var"}.}

\item{full_layer}{whether to output the predictions of all layers. Defaults to \code{FALSE}. Only used when \code{obj} is for DGP and linked (D)GP.}

\item{sample_size}{the number of samples to draw for each given imputation if \code{method = "sampling"}. Defaults to \code{50}.}
}
\value{
\itemize{
\item \strong{(GP case)} If \code{obj} is produced by \code{\link[=train]{train()}} on a object built with \code{\link[=gp]{gp()}}:
\enumerate{
\item if \code{method = "mean_var"}: a named list is returned that contains two matrices named \code{mean} for the predictive means
and \code{var} for the predictive variances. Each matrix has only one column with its rows
corresponding to testing positions (i.e., row of \code{x}).
\item if \code{method = "sampling"}: a matrix is returned whose rows corresponding to testing positions and columns corresponding
to \code{sample_size} number of samples drawn from the predictive distribution of GP.
}
\item \strong{(DGP case)} If \code{obj} is produced by \code{\link[=emulator]{emulator()}}:
\enumerate{
\item if \code{method = "mean_var"} and  \code{full_layer = FALSE}: a named list is returned. The list contains two matrices named \code{mean}
for the predictive means and \code{var} for the predictive variances respectively. Each matrix has its rows corresponding to testing
positions and columns corresponding to DGP global output dimensions (i.e., the number of GP/likelihood nodes in the final layer).
\item if \code{method = "mean_var"} and  \code{full_layer = TRUE}: a named list is returned. The list contains two sub-lists named \code{mean}
for the predictive means and \code{var} for the predictive variances respectively. Each sub-list contains \emph{L} (i.e., the number of layers)
matrices named \verb{layer1, layer2,..., layerL}. Each matrix has its rows corresponding to testing positions and columns corresponding to
output dimensions (i.e., the number of GP nodes from the associated layer and in case of the final layer, it may be the number of the
likelihood nodes).
\item if \code{method = "sampling"} and  \code{full_layer = FALSE}: a named list is returned. The list contains \emph{D} (i.e., the number of GP/likelihood
nodes in the final layer) matrices named \verb{output1, output2,..., outputD}. Each matrix has its rows corresponding to testing positions and
columns corresponding to samples of size: \code{N * sample_size}, where \code{N} is the number of imputations specified in \code{\link[=emulator]{emulator()}}.
\item if \code{method = "sampling"} and  \code{full_layer = TRUE}: a named list is returned. The list contains \emph{L} (i.e., the number of layers)
sub-lists named \verb{layer1, layer2,..., layerL}. Each sub-list represents samples drawn from the GPs/likelihoods in the corresponding layers,
and contains \emph{D} (i.e., the number of GP nodes in the corresponding layer or likelihood nodes in the final layer)
matrices named \verb{output1, output2,..., outputD}. Each matrix gives samples of the output from one of \emph{D} GPs/likelihoods at the
testing positions, and has its rows corresponding to testing positions and columns corresponding to samples
of size: \code{N * sample_size}, where \code{N} is the number of imputations specified in \code{\link[=emulator]{emulator()}}.
}
\item \strong{(Linked (D)GP case)} If \code{obj} is produced by \code{\link[=lgp]{lgp()}}:
\enumerate{
\item if \code{method = "mean_var"} and  \code{full_layer = FALSE}: a named list is returned. The list contains two sub-lists named \code{mean}
for the predictive means and \code{var} for the predictive variances respectively. Each sub-list contains \emph{M} number (same number
of computer models in the final layer of the system) of matrices named \verb{emulator1, emulator2,..., emulatorM}. Each matrix has
its rows corresponding to global testing positions and columns corresponding to output dimensions of the associated emulator
in the final layer.
\item if \code{method = "mean_var"} and  \code{full_layer = TRUE}: a named list is returned. The list contains two sub-lists named \code{mean}
for the predictive means and \code{var} for the predictive variances respectively. Each sub-list contains \emph{L} (i.e., the number of layers
of the emulated system) components named \verb{layer1, layer2,..., layerL}. Each component represents a layer and contains \emph{M} number
(same number of emulators in the corresponding layer of the system) of matrices named \verb{emulator1, emulator2,..., emulatorM}. Each
matrix has its rows corresponding to global testing positions and columns corresponding to output dimensions of the associated
GP/DGP emulator in the corresponding layer.
\item if \code{method = "sampling"} and  \code{full_layer = FALSE}: a named list is returned. The list contains \emph{M} number (same number of emulators
in the final layer of the system) of sub-lists named \verb{emulator1, emulator2,..., emulatorM}. Each sub-list corresponds to an emulator
in the final layer, and contains \emph{D} matrices, named \verb{output1, output2,..., outputD}, that correspond to the output
dimensions of the GP/DGP emulator. Each matrix has its rows corresponding to testing positions and columns corresponding to samples
of size: \code{N * sample_size}, where \code{N} is the number of imputations specified in \code{\link[=emulator]{emulator()}}.
\item if \code{method = "sampling"} and  \code{full_layer = TRUE}: a named list is returned. The list contains \emph{L} (i.e., the number of layers of
the emulated system) sub-lists named \verb{layer1, layer2,..., layerL}. Each sub-list represents a layer and contains \emph{M} number (same
number of emulators in the corresponding layer of the system) of components named \verb{emulator1, emulator2,..., emulatorM}. Each component
corresponds to an emulator in the associated layer, and contains \emph{D} matrices, named \verb{output1, output2,..., outputD}, that correspond to
the output dimensions of the GP/DGP emulator. Each matrix has its rows corresponding to testing positions and columns corresponding to
samples of size: \code{N * sample_size}, where \code{N} is the number of imputations specified in \code{\link[=emulator]{emulator()}}.
}
}
}
\description{
This function implements prediction from GP, DGP, or linked (D)GP emulators.
}
