% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read}
\alias{read}
\title{Load the stored emulator}
\usage{
read(pkl_file)
}
\arguments{
\item{pkl_file}{the path to and the name of the \code{.pkl} file where the emulator is stored.}
}
\value{
A GP, DGP or linked (D)GP emulator S3 class.
}
\description{
This function loads the \code{.pkl} file that stores the emulator.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\examples{
\dontrun{

# See gp(), dgp(), or lgp() for an example.
}
}
