% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_linked_idx}
\alias{set_linked_idx}
\title{Set linked indices}
\usage{
set_linked_idx(object, idx)
}
\arguments{
\item{object}{an instance of the S3 class \code{gp} or \code{dgp}.}

\item{idx}{same as the argument \code{linked_idx} of \code{\link[=gp]{gp()}} and \code{\link[=dgp]{dgp()}}.}
}
\value{
An updated \code{object} with the information of \code{idx} incorporated.
}
\description{
This function adds the linked information to a GP or DGP emulator if the information is not provided
when the emulator is constructed by \code{\link[=gp]{gp()}} or \code{\link[=dgp]{dgp()}}.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\note{
This function is useful when different models are emulated by different teams. Each team can create their (D)GP emulator
even without knowing how different emulators are connected together. When this information is available and
different emulators are collected, the connection information between emulators can then be assigned to
individual emulators with this function.
}
\examples{
\dontrun{

# See lgp() for an example.
}
}
