% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary}
\alias{summary}
\alias{summary.gp}
\alias{summary.dgp}
\alias{summary.lgp}
\title{Visual summary of a constructed GP, DGP, or linked (D)GP emulator}
\usage{
\method{summary}{gp}(object, ...)

\method{summary}{dgp}(object, ...)

\method{summary}{lgp}(object, ...)
}
\arguments{
\item{object}{can be one of the following:
\itemize{
\item the S3 class \code{gp}.
\item the S3 class \code{dgp}.
\item the S3 class \code{lgp}.
}}

\item{...}{N/A.}
}
\value{
An interactive visualization of the emulator as a \code{visNetwork} object.
This visualization is compatible with R Markdown documents and the RStudio Viewer.
The resulting \code{visNetwork} object can be saved as an HTML file using \code{\link[visNetwork:visSave]{visNetwork::visSave()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#updated}{\figure{lifecycle-updated.svg}{options: alt='[Updated]'}}}{\strong{[Updated]}}

This function provides a summary of key information for a GP, DGP, or linked (D)GP emulator
by generating an interactive visualization of the emulator’s structure.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\examples{
\dontrun{

# See gp(), dgp(), or lgp() for an example.
}
}
