% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary}
\alias{summary}
\title{Summary of constructed GP, DGP, and linked (D)GP structures}
\usage{
summary(obj)
}
\arguments{
\item{obj}{can be one of the following:
* the object produced by \code{\link[=kernel]{kernel()}}.
* the object produced by \code{\link[=gp]{gp()}} (either before or after applying \code{\link[=train]{train()}}).
* the object produced by \code{\link[=dgp]{dgp()}} before applying \code{\link[=train]{train()}}.
* the object produced by \code{\link[=emulator]{emulator()}}.
* the object produced by \code{\link[=lgp]{lgp()}}.}
}
\value{
A table summarizing key information contained in \code{obj}.
}
\description{
This function summarizes key information of GP, DGP, and linked (D)GP structures.
}
\details{
See examples in tutorials at \url{https://mingdeyu.github.io/dgpsi-R}.
}
