% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trace_plot}
\alias{trace_plot}
\title{Plot of DGP model parameter traces}
\usage{
trace_plot(object, layer = NULL, node = 1)
}
\arguments{
\item{object}{an instance of the \code{dgp} class.}

\item{layer}{the index of a layer. Defaults to \code{NULL} for the final layer.}

\item{node}{the index of a GP node in the layer specified by \code{layer}. Defaults to \code{1} for the first GP node in the
corresponding layer.}
}
\value{
A \code{ggplot} object.
}
\description{
This function plots the traces of model parameters of a chosen GP node
in a DGP emulator.
}
\details{
See further examples and tutorials at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\examples{
\dontrun{

# See dgp() for an example.
}
}
