% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write}
\alias{write}
\title{Save the constructed emulator}
\usage{
write(object, pkl_file)
}
\arguments{
\item{object}{an instance of the S3 class \code{gp}, \code{dgp}, or \code{lgp}.}

\item{pkl_file}{the path to and the name of the \code{.pkl} file to which
the emulator \code{object} is saved.}
}
\description{
This function saves the constructed emulator to a \code{.pkl} file.
}
\details{
See examples in Articles at \url{https://mingdeyu.github.io/dgpsi-R/}.
}
\note{
Since the constructed emulators are Python objects, \code{\link[=save]{save()}} from R will not work as it is only for R objects.
}
